PJS: The Gaussians have returned
git-svn-id: https://projectm.svn.sourceforge.net/svnroot/projectm/trunk@842 6778bc44-b910-0410-a7a0-be141de4315d

+++ src/projectM-engine/projectM.hpp
@@ -216,7 +216,7 @@ public:
   }
 private:
 	
-  int sampledPresetDuration();
+  double sampledPresetDuration();
   BeatDetect * beatDetect;
   Renderer *renderer;
   Settings _settings;
